requirements.txt
django
django-extensions
djangorestframework
requests
spotipy
Pillow
fuzzywuzzy
python-Levenshtein
django-import-export
social-auth-app-django
